requirements.txt
attrs
beautifulsoup4
black
bs4
certifi
charset-normalizer
click
cycler
et-xmlfile
exceptiongroup
fonttools
h11
idna
importlib-metadata
kiwisolver
matplotlib
mypy-extensions
numpy
openpyxl
outcome
packaging
pandas
pathspec
Pillow
platformdirs
PyMySQL
pyparsing
PySocks
python-dateutil
python-decouple
pytz
requests
seaborn
selenium
six
sniffio
sortedcontainers
soupsieve
tomli
trio
trio-websocket
typed-ast
typing_extensions
urllib3
wsproto
zipp
google-generativeai
streamlit
python-dotenv
